añadi cambioas en las clases PLANBASICO, PLANESTANDAR Y PLAN PREMIUM, SOLO LES AÑADI UN (double) EN LA LINEA DE RETURN DEL METODO ADICIONAL A ESO, AÑADI LA CLASE PLATAFORMASTREAMING

+++ src/Planbasico.java
@@ -1,5 +1,5 @@
 public class Planbasico implements PlanSuscripcion {
     public double calcularCosto(int meses){
-        return meses*5;
+        return (double) meses*5;
     }
 }
